+++ verifier/src/main/java/com/aws/jverify/verifier/dafny/Wrapper.java
@@ -1,0 +1,17 @@
+package com.aws.jverify.verifier.dafny;
+
+public class Wrapper<T> {
+    T value;
+
+    public Wrapper(T value) {
+        this.value = value;
+    }
+
+    public T getValue() {
+        return value;
+    }
+
+    public void setValue(T value) {
+        this.value = value;
+    }
+}
